+++ apps/openapi-demo/app/const.ts
@@ -1,2 +1,0 @@
-// TODO: IN THE FUTURE BETTER cuz should also work without prefix
-export const DOCS_PREFIX = "/docs/api"