+++ packages/xyd-framework/packages/react/components/index.tsx
@@ -158,7 +158,6 @@ export function FwSidebarGroups(props: FwSidebarGroupsProps) {
         return group
     })
 
-
     // TODO: better API for elements like logo search
     return <FwSidebarGroupContext
         initialActiveItems={initialActiveItems}