+++ packages/xyd-openapi/__tests__/oapSchemaToReferences.test.ts
@@ -13,32 +13,48 @@ const tests: {
     options?: uniformOasOptions
 }[] = [
     {
-        name: "1.basic",
-        description: "Basic OpenAPI API example",
-    },
-    {
-        name: "2.more",
-        description: "More OpenAPI API example",
-    },
-    {
-        name: "3.multiple-responses",
-        description: "Multiple responses OpenAPI API example",
-    },
-    {
-        name: "5.xdocs.codeLanguages",
-        description: "x-docs OpenAPI API codeLanguages example",
-    },
-    {
-        name: "5.xdocs.sidebar",
-        description: "x-docs OpenAPI API sidebar example",
+        name: "-3.random",
+        // url: "https://raw.githubusercontent.com/bump-sh-examples/train-travel-api/main/openapi.yaml",
+        // url: "https://raw.githubusercontent.com/digitalocean/openapi/main/specification/DigitalOcean-public.v2.yaml",
+        // url: "https://raw.githubusercontent.com/github/rest-api-description/main/descriptions/ghes-3.0/ghes-3.0.json",
+        // url: "https://raw.githubusercontent.com/box/box-openapi/main/openapi.json",
+        // url: "https://api.apis.guru/v2/specs/nytimes.com/article_search/1.0.0/openapi.yaml",
+        url: "https://developers.intercom.com/_spec/docs/references/@2.11/rest-api/api.intercom.io.json",
+        // url: "https://raw.githubusercontent.com/livesession/livesession-openapi/refs/heads/master/openapi.yaml",
+        // url: "https://openapi.vercel.sh",
+        description: "",
         plugins: [
             uniformPluginXDocsSidebar
         ]
     },
-    {
-        name: "6.codeSamples",
-        description: "x-codeSamples OpenAPI API example",
-    },
+
+    // {
+    //     name: "1.basic",
+    //     description: "Basic OpenAPI API example",
+    // },
+    // {
+    //     name: "2.more",
+    //     description: "More OpenAPI API example",
+    // },
+    // {
+    //     name: "3.multiple-responses",
+    //     description: "Multiple responses OpenAPI API example",
+    // },
+    // {
+    //     name: "5.xdocs.codeLanguages",
+    //     description: "x-docs OpenAPI API codeLanguages example",
+    // },
+    // {
+    //     name: "5.xdocs.sidebar",
+    //     description: "x-docs OpenAPI API sidebar example",
+    //     plugins: [
+    //         uniformPluginXDocsSidebar
+    //     ]
+    // },
+    // {
+    //     name: "6.codeSamples",
+    //     description: "x-codeSamples OpenAPI API example",
+    // },
 ]
 
 describe("oapSchemaToReferences", {timeout: 15000}, () => {
@@ -64,16 +80,3 @@ describe("oapSchemaToReferences", {timeout: 15000}, () => {
 //     //    ]
 //     // }
 // },
-// {
-//     name: "-3.random",
-//     // url: "https://raw.githubusercontent.com/bump-sh-examples/train-travel-api/main/openapi.yaml",
-//     // url: "https://raw.githubusercontent.com/digitalocean/openapi/main/specification/DigitalOcean-public.v2.yaml",
-//     // url: "https://raw.githubusercontent.com/github/rest-api-description/main/descriptions/ghes-3.0/ghes-3.0.json",
-//     // url: "https://raw.githubusercontent.com/box/box-openapi/main/openapi.json",
-//     // url: "https://api.apis.guru/v2/specs/nytimes.com/article_search/1.0.0/openapi.yaml",
-//     // url: "https://developers.intercom.com/_spec/docs/references/@2.11/rest-api/api.intercom.io.json",
-//     description: "",
-//     plugins: [
-//         uniformPluginXDocsSidebar
-//     ]
-// }