+++ packages/xyd-atlas/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @xyd-js/atlas
 
+## 0.1.0-xyd.99
+
+### Patch Changes
+
+- update all packages
+- Updated dependencies
+  - @xyd-js/components@0.1.0-xyd.94
+  - @xyd-js/core@0.1.0-xyd.96
+  - @xyd-js/framework@0.1.0-xyd.115
+  - @xyd-js/plugins@0.1.0-xyd.83
+  - @xyd-js/uniform@0.1.0-xyd.98
+
 ## 0.1.0-xyd.98
 
 ### Patch Changes