+++ packages/xyd-plugins/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @xyd-js/plugins
 
+## 0.1.0-xyd.83
+
+### Patch Changes
+
+- update all packages
+- Updated dependencies
+  - @xyd-js/framework@0.1.0-xyd.115
+  - @xyd-js/uniform@0.1.0-xyd.98
+
 ## 0.1.0-xyd.82
 
 ### Patch Changes