+++ packages/xyd-storybook/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @xyd-js/storybook
 
+## 0.0.1-xyd.93
+
+### Patch Changes
+
+- update all packages
+- Updated dependencies
+  - @xyd-js/atlas@0.1.0-xyd.99
+  - @xyd-js/components@0.1.0-xyd.94
+  - @xyd-js/content@0.1.0-xyd.97
+  - @xyd-js/theme-cosmo@0.1.0-xyd.83
+  - @xyd-js/theme-opener@0.1.0-xyd.83
+  - @xyd-js/theme-picasso@0.1.0-xyd.83
+  - @xyd-js/theme-poetry@0.1.0-xyd.110
+  - @xyd-js/ui@0.1.0-xyd.99
+
 ## 0.0.1-xyd.92
 
 ### Patch Changes