+++ packages/xyd-ui/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @xyd-js/ui
 
+## 0.1.0-xyd.99
+
+### Patch Changes
+
+- update all packages
+- Updated dependencies
+  - @xyd-js/components@0.1.0-xyd.94
+
 ## 0.1.0-xyd.98
 
 ### Patch Changes