+++ packages/xyd-uniform/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @xyd-js/uniform
 
+## 0.1.0-xyd.98
+
+### Patch Changes
+
+- update all packages
+- Updated dependencies
+  - @xyd-js/core@0.1.0-xyd.96
+
 ## 0.1.0-xyd.97
 
 ### Patch Changes